ints: Pull types into independent files
In anticipation of automatically generating these definitions, pull the
definitions of the various integer types into their own files as well as
their tests.

This will make it easier to review the changes when each of these files
is generated independently and automatically.
