+++ atomic_test.go
@@ -29,158 +29,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestInt32(t *testing.T) {
-	atom := NewInt32(42)
-
-	require.Equal(t, int32(42), atom.Load(), "Load didn't work.")
-	require.Equal(t, int32(46), atom.Add(4), "Add didn't work.")
-	require.Equal(t, int32(44), atom.Sub(2), "Sub didn't work.")
-	require.Equal(t, int32(45), atom.Inc(), "Inc didn't work.")
-	require.Equal(t, int32(44), atom.Dec(), "Dec didn't work.")
-
-	require.True(t, atom.CAS(44, 0), "CAS didn't report a swap.")
-	require.Equal(t, int32(0), atom.Load(), "CAS didn't set the correct value.")
-
-	require.Equal(t, int32(0), atom.Swap(1), "Swap didn't return the old value.")
-	require.Equal(t, int32(1), atom.Load(), "Swap didn't set the correct value.")
-
-	atom.Store(42)
-	require.Equal(t, int32(42), atom.Load(), "Store didn't set the correct value.")
-
-	t.Run("JSON/Marshal", func(t *testing.T) {
-		bytes, err := json.Marshal(atom)
-		require.NoError(t, err, "json.Marshal errored unexpectedly.")
-		require.Equal(t, []byte("42"), bytes, "json.Marshal encoded the wrong bytes.")
-	})
-
-	t.Run("JSON/Unmarshal", func(t *testing.T) {
-		err := json.Unmarshal([]byte("40"), &atom)
-		require.NoError(t, err, "json.Unmarshal errored unexpectedly.")
-		require.Equal(t, int32(40), atom.Load(), "json.Unmarshal didn't set the correct value.")
-	})
-
-	t.Run("JSON/Unmarshal/Error", func(t *testing.T) {
-		err := json.Unmarshal([]byte(`"40"`), &atom)
-		require.Error(t, err, "json.Unmarshal didn't error as expected.")
-		assertErrorJSONUnmarshalType(t, err,
-			"json.Unmarshal failed with unexpected error %v, want UnmarshalTypeError.", err)
-	})
-}
-
-func TestInt64(t *testing.T) {
-	atom := NewInt64(42)
-
-	require.Equal(t, int64(42), atom.Load(), "Load didn't work.")
-	require.Equal(t, int64(46), atom.Add(4), "Add didn't work.")
-	require.Equal(t, int64(44), atom.Sub(2), "Sub didn't work.")
-	require.Equal(t, int64(45), atom.Inc(), "Inc didn't work.")
-	require.Equal(t, int64(44), atom.Dec(), "Dec didn't work.")
-
-	require.True(t, atom.CAS(44, 0), "CAS didn't report a swap.")
-	require.Equal(t, int64(0), atom.Load(), "CAS didn't set the correct value.")
-
-	require.Equal(t, int64(0), atom.Swap(1), "Swap didn't return the old value.")
-	require.Equal(t, int64(1), atom.Load(), "Swap didn't set the correct value.")
-
-	atom.Store(42)
-	require.Equal(t, int64(42), atom.Load(), "Store didn't set the correct value.")
-
-	t.Run("JSON/Marshal", func(t *testing.T) {
-		bytes, err := json.Marshal(atom)
-		require.NoError(t, err, "json.Marshal errored unexpectedly.")
-		require.Equal(t, []byte("42"), bytes, "json.Marshal encoded the wrong bytes.")
-	})
-
-	t.Run("JSON/Unmarshal", func(t *testing.T) {
-		err := json.Unmarshal([]byte("40"), &atom)
-		require.NoError(t, err, "json.Unmarshal errored unexpectedly.")
-		require.Equal(t, int64(40), atom.Load(), "json.Unmarshal didn't set the correct value.")
-	})
-
-	t.Run("JSON/Unmarshal/Error", func(t *testing.T) {
-		err := json.Unmarshal([]byte(`"40"`), &atom)
-		require.Error(t, err, "json.Unmarshal didn't error as expected.")
-		assertErrorJSONUnmarshalType(t, err,
-			"json.Unmarshal failed with unexpected error %v, want UnmarshalTypeError.", err)
-	})
-}
-
-func TestUint32(t *testing.T) {
-	atom := NewUint32(42)
-
-	require.Equal(t, uint32(42), atom.Load(), "Load didn't work.")
-	require.Equal(t, uint32(46), atom.Add(4), "Add didn't work.")
-	require.Equal(t, uint32(44), atom.Sub(2), "Sub didn't work.")
-	require.Equal(t, uint32(45), atom.Inc(), "Inc didn't work.")
-	require.Equal(t, uint32(44), atom.Dec(), "Dec didn't work.")
-
-	require.True(t, atom.CAS(44, 0), "CAS didn't report a swap.")
-	require.Equal(t, uint32(0), atom.Load(), "CAS didn't set the correct value.")
-
-	require.Equal(t, uint32(0), atom.Swap(1), "Swap didn't return the old value.")
-	require.Equal(t, uint32(1), atom.Load(), "Swap didn't set the correct value.")
-
-	atom.Store(42)
-	require.Equal(t, uint32(42), atom.Load(), "Store didn't set the correct value.")
-
-	t.Run("JSON/Marshal", func(t *testing.T) {
-		bytes, err := json.Marshal(atom)
-		require.NoError(t, err, "json.Marshal errored unexpectedly.")
-		require.Equal(t, []byte("42"), bytes, "json.Marshal encoded the wrong bytes.")
-	})
-
-	t.Run("JSON/Unmarshal", func(t *testing.T) {
-		err := json.Unmarshal([]byte("40"), &atom)
-		require.NoError(t, err, "json.Unmarshal errored unexpectedly.")
-		require.Equal(t, uint32(40), atom.Load(), "json.Unmarshal didn't set the correct value.")
-	})
-
-	t.Run("JSON/Unmarshal/Error", func(t *testing.T) {
-		err := json.Unmarshal([]byte(`"40"`), &atom)
-		require.Error(t, err, "json.Unmarshal didn't error as expected.")
-		assertErrorJSONUnmarshalType(t, err,
-			"json.Unmarshal failed with unexpected error %v, want UnmarshalTypeError.", err)
-	})
-}
-
-func TestUint64(t *testing.T) {
-	atom := NewUint64(42)
-
-	require.Equal(t, uint64(42), atom.Load(), "Load didn't work.")
-	require.Equal(t, uint64(46), atom.Add(4), "Add didn't work.")
-	require.Equal(t, uint64(44), atom.Sub(2), "Sub didn't work.")
-	require.Equal(t, uint64(45), atom.Inc(), "Inc didn't work.")
-	require.Equal(t, uint64(44), atom.Dec(), "Dec didn't work.")
-
-	require.True(t, atom.CAS(44, 0), "CAS didn't report a swap.")
-	require.Equal(t, uint64(0), atom.Load(), "CAS didn't set the correct value.")
-
-	require.Equal(t, uint64(0), atom.Swap(1), "Swap didn't return the old value.")
-	require.Equal(t, uint64(1), atom.Load(), "Swap didn't set the correct value.")
-
-	atom.Store(42)
-	require.Equal(t, uint64(42), atom.Load(), "Store didn't set the correct value.")
-
-	t.Run("JSON/Marshal", func(t *testing.T) {
-		bytes, err := json.Marshal(atom)
-		require.NoError(t, err, "json.Marshal errored unexpectedly.")
-		require.Equal(t, []byte("42"), bytes, "json.Marshal encoded the wrong bytes.")
-	})
-
-	t.Run("JSON/Unmarshal", func(t *testing.T) {
-		err := json.Unmarshal([]byte("40"), &atom)
-		require.NoError(t, err, "json.Unmarshal errored unexpectedly.")
-		require.Equal(t, uint64(40), atom.Load(), "json.Unmarshal didn't set the correct value.")
-	})
-
-	t.Run("JSON/Unmarshal/Error", func(t *testing.T) {
-		err := json.Unmarshal([]byte(`"40"`), &atom)
-		require.Error(t, err, "json.Unmarshal didn't error as expected.")
-		assertErrorJSONUnmarshalType(t, err,
-			"json.Unmarshal failed with unexpected error %v, want UnmarshalTypeError.", err)
-	})
-}
-
 func TestBool(t *testing.T) {
 	atom := NewBool(false)
 	require.False(t, atom.Toggle(), "Expected Toggle to return previous value.")
@@ -253,7 +101,7 @@ func TestFloat64(t *testing.T) {
 	})
 
 	t.Run("JSON/Unmarshal/Error", func(t *testing.T) {
-		err := json.Unmarshal([]byte(`"40.5"`), &atom)
+		err := json.Unmarshal([]byte("\"40.5\""), &atom)
 		require.Error(t, err, "json.Unmarshal didn't error as expected.")
 		assertErrorJSONUnmarshalType(t, err,
 			"json.Unmarshal failed with unexpected error %v, want UnmarshalTypeError.", err)
@@ -290,7 +138,7 @@ func TestDuration(t *testing.T) {
 	})
 
 	t.Run("JSON/Unmarshal/Error", func(t *testing.T) {
-		err := json.Unmarshal([]byte(`"1000000000"`), &atom)
+		err := json.Unmarshal([]byte("\"1000000000\""), &atom)
 		require.Error(t, err, "json.Unmarshal didn't error as expected.")
 		assertErrorJSONUnmarshalType(t, err,
 			"json.Unmarshal failed with unexpected error %v, want UnmarshalTypeError.", err)
